4.5.0

+++ src/pages/workflow/WorkflowList/components/DeleteFlowBtn/index.jsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import { DeleteReconfirm } from 'ming-ui';
+import Confirm from 'ming-ui/components/Dialog/Confirm';
 import process from '../../../api/process';
 
 export default class DeleteFlowBtn extends Component {
@@ -9,10 +9,11 @@ export default class DeleteFlowBtn extends Component {
   delFlow = () => {
     const { item, callback } = this.props;
 
-    DeleteReconfirm({
+    Confirm({
       title: _l('删除工作流“%0”', item.name),
-      description: _l('工作流将被彻底删除，且无法恢复。请确认您要执行此操作'),
-      data: [{ text: _l('我确定执行此操作'), value: true }],
+      description: _l('工作流将被删除，请确认执行此操作'),
+      okText:_l('删除'),
+      buttonType: 'danger',
       onOk: () => {
         process.deleteProcess({ processId: item.id }).then(res => {
           if (res) {
